Split out account validation from services
